Struct diesel_logger::LoggingConnection

pub struct LoggingConnection<C: Connection>(_);

Wraps a diesel Connection to time and log each query using the configured logger for the log crate.

Currently, this produces a debug log on every query, an info on queries that take longer than 1 second, and a warning on queries that take longer than 5 seconds. These thresholds will be configurable in a future version.

Methods

impl<C: Connection> LoggingConnection<C>

pub fn new(conn: C) -> Self